refactor: State에서 이긴 상태의 수익만 계산하는 것이 아닌 패배, 무승부일때도 수익 계산하도록 변경

+++ src/main/java/blackjack/domain/participants/Dealer.java
@@ -29,7 +29,7 @@ public class Dealer {
         return cards.calculateMaxScore();
     }
 
-    public int calculateProfit(Players players) {
+    public double calculateProfit(Players players) {
         return -1 * players.calculateTotalProfit(cards);
     }
