+++ src/main/java/blackjack/domain/participants/Profit.java
@@ -3,10 +3,10 @@ package blackjack.domain.participants;
 import java.util.Map;
 
 public record Profit(
-        int dealerProfit,
-        Map<Player, Integer> playerProfit
+        double dealerProfit,
+        Map<Player, Double> playerProfit
 ) {
-    public Profit(int dealerProfit, Map<Player, Integer> playerProfit) {
+    public Profit(double dealerProfit, Map<Player, Double> playerProfit) {
         this.dealerProfit = dealerProfit;
         this.playerProfit = playerProfit;
     }
